# lib/api-zod/src/generated/types/assetCuez.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Employee Master API specification
 * OpenAPI spec version: 0.1.0
 */

export interface AssetCuez {
  id: number;
  employeeId: string;
  firstName: string;
  lastName?: string;
  contactNumber?: string;
  activationStatus?: string;
  createdAt: string;
}