# lib/api-zod/src/generated/types/createAcronisRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Employee Master API specification
 * OpenAPI spec version: 0.1.0
 */

export interface CreateAcronisRequest {
  employeeId: string;
  employeeName: string;
  department?: string;
  designation?: string;
  email?: string;
  hostname?: string;
  backupPlan?: string;
  monthlyFullBackupDate?: string;
  backupPlanName?: string;
  backupTime?: string;
  incrementalBackup?: string;
  cpuPriority?: string;
  installationStatus?: string;
  policyAssign?: string;
  backupStatus?: string;
  leftEncryptionPassword?: string;
}