# lib/api-zod/src/generated/types/jiraUser.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Employee Master API specification
 * OpenAPI spec version: 0.1.0
 */

export interface JiraUser {
  id: number;
  employeeId: string;
  username: string;
  email?: string;
  userStatus?: string;
  addedToOrganization?: string;
  createdAt: string;
}