# lib/api-zod/src/generated/types/vpnUser.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Employee Master API specification
 * OpenAPI spec version: 0.1.0
 */

export interface VpnUser {
  id: number;
  employeeId: string;
  username: string;
  department?: string;
  designation?: string;
  userId?: string;
  password?: string;
  status?: string;
  createdAt: string;
}